+++ RESearch.h
@@ -253,7 +253,7 @@ enum {
 void ReadRegistry();
 void WriteRegistry();
 
-void PrepareBMHSearch(const char *String,int StringLength,int nPattern = 0);
+void PrepareBMHSearch(const char *String,int StringLength,size_t nPattern = 0);
 BOOL PreparePattern(pcre **Pattern,pcre_extra **PatternExtra,string &Text,int CaseSensitive,BOOL bUTF8=FALSE);
 char *CreateReplaceString(const char *Matched,int *Match,int Count,const char *Replace,const char *EOL,int *Numbers,int &ResultLength);
